Extract HasCursor trait from campaign model.
